app/requirements.txt
pylint
pytest
flask
pandas
psycopg2-binary
python-dotenv